G5

+++ m2/s3/g5/script.js
@@ -4,6 +4,7 @@ let brandInput = document.getElementById("brand");
 let priceInput = document.getElementById("price");
 let imageInput = document.getElementById("img");
 let rowReference = document.getElementById("product-container");
+let reset = document.getElementById("reset-button");
 
 let saveButton = document.getElementById("save-button");
 let deleteButton = document.getElementById("delete-button");
@@ -39,8 +40,9 @@ let getProduct = function () {
     })
     .then((data) => {
       data.forEach((event) => {
+        console.log(event);
         let colTemplate = `
-              <div class="col-12 col-md-3">
+              <div class="col-12 ">
                 <div class="card shadow-lg border-0">
                 <img class="w-100" style="height:300px" src="${event.imageUrl}" alt="">
                   <div class="card-body">
@@ -49,10 +51,12 @@ let getProduct = function () {
                     <p class="card-text">
                       ${event.brand}
                     </p>
-                    <p>${event.description}</p>
+                    <p style="white-space: nowrap;
+                    overflow: hidden;
+                    text-overflow: ellipsis;">${event.description}</p>
                     <p>${event.price}€</p>
-                    <a href="backoffice.html?eventId=${event._id}" class="btn btn-primary ms-1">MODIFICA</a>
-                    <a href="details.html?eventId=${event._id}" class="btn btn-primary" data-toggle="modal" data-target="#examp">SCOPRI DI PIU</a>
+                    <a href="backoffice.html?eventId=${event._id}" class="btn btn-outline-secondary ms-1">MODIFICA</a>
+                    <a href="details.html?eventId=${event._id}" class="btn btn-outline-secondary" data-toggle="modal" data-target="#examp">SCOPRI DI PIU</a>
                   </div>
                 </div>
               </div>
@@ -117,36 +121,35 @@ let addressBarContent = new URLSearchParams(window.location.search);
 
 let eventId = addressBarContent.get("eventId");
 
-deleteButton.addEventListener("click", () => {
-  fetch(
-    eventId
-      ? "https://striveschool-api.herokuapp.com/api/product/" + eventId
-      : "https://striveschool-api.herokuapp.com/api/product/",
-    {
-      method: "DELETE",
-      headers: {
-        Authorization:
-          "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJfaWQiOiI2NDVkZTRjNzg4Zjc0MDAwMTQyODc0MzQiLCJpYXQiOjE2ODM4NzUwMTUsImV4cCI6MTY4NTA4NDYxNX0.UhZMEpCcZJmZtlPfSyEdcq34wjasu6-Wo05gDtHuyTA",
-        "Content-Type": "application/json",
-      },
-    }
-  )
-    .then((res) => {
-      if (res.ok) {
-        alert("eliminazione completata con successo");
-        location.assign("homepage.html");
-      } else {
-        throw new Error("Problema nell'eliminazione dell'evento");
+if (eventId) {
+  deleteButton.addEventListener("click", () => {
+    fetch(
+      eventId
+        ? "https://striveschool-api.herokuapp.com/api/product/" + eventId
+        : "https://striveschool-api.herokuapp.com/api/product/",
+      {
+        method: "DELETE",
+        headers: {
+          Authorization:
+            "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJfaWQiOiI2NDVkZTRjNzg4Zjc0MDAwMTQyODc0MzQiLCJpYXQiOjE2ODM4NzUwMTUsImV4cCI6MTY4NTA4NDYxNX0.UhZMEpCcZJmZtlPfSyEdcq34wjasu6-Wo05gDtHuyTA",
+          "Content-Type": "application/json",
+        },
       }
-    })
-    .catch((err) => {
-      console.log(err);
-    });
-});
+    )
+      .then((res) => {
+        if (res.ok) {
+          alert("eliminazione completata con successo");
+          location.assign("homepage.html");
+        } else {
+          throw new Error("Problema nell'eliminazione dell'evento");
+        }
+      })
+      .catch((err) => {
+        console.log(err);
+      });
+  });
 
-/* fetch("https://striveschool-api.herokuapp.com/api/product/", {
-    method: "POST",
-    body: JSON.stringify(new Product()),
+  fetch("https://striveschool-api.herokuapp.com/api/product/" + eventId, {
     headers: {
       Authorization:
         "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJfaWQiOiI2NDVkZTRjNzg4Zjc0MDAwMTQyODc0MzQiLCJpYXQiOjE2ODM4NzUwMTUsImV4cCI6MTY4NTA4NDYxNX0.UhZMEpCcZJmZtlPfSyEdcq34wjasu6-Wo05gDtHuyTA",
@@ -154,12 +157,59 @@ deleteButton.addEventListener("click", () => {
     },
   })
     .then((res) => {
-      console.log(res);
+      if (res.ok) {
+        return res.json();
+      } else {
+        throw new Error("Errore nel recupero dell'evento");
+      }
     })
-    .then((product) => {
-      console.log(product);
+    .then((event) => {
+      console.log(
+        "DATI DEL SINGOLO EVENTO, RECUPERATO TRAMITE GET SINGOLA",
+        event
+      );
+      // ripopoliamo il form, finalmente
+      nameInput.value = event.name;
+      descriptionInput.value = event.description;
+      brandInput.value = event.brand;
+      price.value = event.price;
+      imageInput.value = event.imageUrl; // trucchetto per ripopolare correttamente un input date
     })
     .catch((error) => {
       console.log(error);
     });
-});*/
+} else {
+  // modalità CREAZIONE
+}
+
+reset.addEventListener("click", () => {
+  fetch("https://striveschool-api.herokuapp.com/api/product/" + eventId, {
+    headers: {
+      Authorization:
+        "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJfaWQiOiI2NDVkZTRjNzg4Zjc0MDAwMTQyODc0MzQiLCJpYXQiOjE2ODM4NzUwMTUsImV4cCI6MTY4NTA4NDYxNX0.UhZMEpCcZJmZtlPfSyEdcq34wjasu6-Wo05gDtHuyTA",
+      "Content-Type": "application/json",
+    },
+  })
+    .then((res) => {
+      if (res.ok) {
+        return res.json();
+      } else {
+        throw new Error("Errore nel recupero dell'evento");
+      }
+    })
+    .then((event) => {
+      console.log(
+        "DATI DEL SINGOLO EVENTO, RECUPERATO TRAMITE GET SINGOLA",
+        event
+      );
+      // ripopoliamo il form, finalmente
+      nameInput.value = " ";
+      descriptionInput.value = " ";
+      brandInput.value = " ";
+      price.value = " ";
+      imageInput.value = " "; // trucchetto per ripopolare correttamente un input date
+    })
+    .catch((error) => {
+      console.log(error);
+    });
+});
